requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-celery-beat==2.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-import-export==2.3.0
django-js-asset==1.2.2
django-timezone-field==4.0
et-xmlfile==1.0.1
jdcal==1.4.1
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
PyMySQL==0.10.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
